refactor(code): separate findCode function to codeService
